Pass lang to getSiteConfig via param instead of x-current-language header
