--- site/src/app/[lang]/sitemap.ts
@@ -1,71 +1,0 @@
-import { gql } from "@comet/cms-site";
-import { getSiteConfig } from "@src/config";
-import { createGraphQLFetch } from "@src/util/graphQLClient";
-import { MetadataRoute } from "next";
-
-import { GQLPrebuildPageDataListSitemapQuery, GQLPrebuildPageDataListSitemapQueryVariables } from "./sitemap.generated";
-
-export const dynamic = "force-dynamic"; // don't generate at build time
-
-export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
-    const sitemap: MetadataRoute.Sitemap = [];
-    const siteConfig = await getSiteConfig();
-    const scope = siteConfig.contentScope;
-
-    const graphqlFetch = createGraphQLFetch();
-
-    let totalCount = 0;
-    let currentCount = 0;
-
-    do {
-        const { paginatedPageTreeNodes } = await graphqlFetch<GQLPrebuildPageDataListSitemapQuery, GQLPrebuildPageDataListSitemapQueryVariables>(
-            pageDataListQuery,
-            {
-                scope,
-                offset: currentCount,
-                limit: 50,
-            },
-        );
-        totalCount = paginatedPageTreeNodes.totalCount;
-        currentCount += paginatedPageTreeNodes.nodes.length;
-
-        for (const pageTreeNode of paginatedPageTreeNodes.nodes) {
-            if (pageTreeNode) {
-                const path: string = pageTreeNode.path;
-
-                if (path && pageTreeNode.document?.__typename === "Page") {
-                    const seoBlock = pageTreeNode.document.seo;
-                    if (!seoBlock.noIndex) {
-                        sitemap.push({
-                            url: siteConfig.url + pageTreeNode.path,
-                            priority: Number(seoBlock.priority.replace("_", ".")),
-                            changeFrequency: seoBlock.changeFrequency,
-                            lastModified: pageTreeNode.document.updatedAt,
-                        });
-                    }
-                }
-            }
-        }
-    } while (totalCount > currentCount);
-
-    return sitemap;
-}
-
-const pageDataListQuery = gql`
-    query PrebuildPageDataListSitemap($scope: PageTreeNodeScopeInput!, $offset: Int, $limit: Int) {
-        paginatedPageTreeNodes(scope: $scope, offset: $offset, limit: $limit) {
-            nodes {
-                id
-                path
-                document {
-                    __typename
-                    ... on Page {
-                        updatedAt
-                        seo
-                    }
-                }
-            }
-            totalCount
-        }
-    }
-`;
