+++ site/src/app/sitemap.ts
@@ -1,0 +1,78 @@
+import { gql } from "@comet/cms-site";
+import { getSiteConfigs } from "@src/siteConfigs";
+import { createGraphQLFetch } from "@src/util/graphQLClient";
+import { MetadataRoute } from "next";
+
+import { GQLPrebuildPageDataListSitemapQuery, GQLPrebuildPageDataListSitemapQueryVariables } from "./sitemap.generated";
+
+export const dynamic = "force-dynamic"; // don't generate at build time
+
+const getUrlFromDomain = (domain: string): string => {
+    return domain.includes("localhost") ? `http://${domain}` : `https://${domain}`;
+};
+
+export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
+    const sitemap: MetadataRoute.Sitemap = [];
+    const graphqlFetch = createGraphQLFetch();
+
+    const siteConfigs = getSiteConfigs();
+
+    for (const siteConfig of siteConfigs) {
+        const scope = siteConfig.contentScope;
+
+        let totalCount = 0;
+        let currentCount = 0;
+
+        do {
+            const { paginatedPageTreeNodes } = await graphqlFetch<GQLPrebuildPageDataListSitemapQuery, GQLPrebuildPageDataListSitemapQueryVariables>(
+                pageDataListQuery,
+                {
+                    scope,
+                    offset: currentCount,
+                    limit: 50,
+                },
+            );
+            totalCount = paginatedPageTreeNodes.totalCount;
+            currentCount += paginatedPageTreeNodes.nodes.length;
+
+            for (const pageTreeNode of paginatedPageTreeNodes.nodes) {
+                if (pageTreeNode) {
+                    const path: string = pageTreeNode.path;
+
+                    if (path && pageTreeNode.document?.__typename === "Page") {
+                        const seoBlock = pageTreeNode.document.seo;
+                        if (!seoBlock.noIndex) {
+                            sitemap.push({
+                                url: `${getUrlFromDomain(siteConfig.domains.main)}/${scope.language}${pageTreeNode.path}`, // TODO support multiple site domains
+                                priority: Number(seoBlock.priority.replace("_", ".")),
+                                changeFrequency: seoBlock.changeFrequency,
+                                lastModified: pageTreeNode.document.updatedAt,
+                            });
+                        }
+                    }
+                }
+            }
+        } while (totalCount > currentCount);
+    }
+
+    return sitemap;
+}
+
+const pageDataListQuery = gql`
+    query PrebuildPageDataListSitemap($scope: PageTreeNodeScopeInput!, $offset: Int, $limit: Int) {
+        paginatedPageTreeNodes(scope: $scope, offset: $offset, limit: $limit) {
+            nodes {
+                id
+                path
+                document {
+                    __typename
+                    ... on Page {
+                        updatedAt
+                        seo
+                    }
+                }
+            }
+            totalCount
+        }
+    }
+`;
